requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pymongo==4.6.1
cassandra-driver==3.29.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
faker==22.6.0 